refactor: modified inventory handler; change monet types to decimal; use inventory and refactor order service

+++ OrderProcessing/GooglePaymentService.cs
@@ -2,7 +2,7 @@
 
 public class GooglePaymentService : IPaymentService
 {
-  public bool ProcessPayment(Customer customer, double amount)
+  public bool ProcessPayment(Customer customer, decimal amount)
   {
     Console.WriteLine($"Processing payment for {customer.Name} of amount {amount}");
